New homepage and a web demo via RacketScript
1. Adds a new homepage. It can be built and pushed to GitHub pages by
   running:

   ```bash
   build/push-gh-pages
   ```

   Demo: https://glebm.github.io/waxeye/index.html

2. Adds a web demo!
   It runs directly in the browser.
   The waxeye compiler is compiled using RacketScript.

   Demo: https://glebm.github.io/waxeye/demo.html

Minimal changes were necessary on the waxeye side to make it work with
RacketScript. Incidentally, these changes also reduced the size of the
final executable by ~50%.

On the RacketScript side, I had to fix a few bugs in RacketScript:
https://github.com/vishesh/racketscript/pull/73

Once the RacketScript PR is merged and a new version of RacketScript
is released, this PR can be merged. For now, I just want your feedback.

@jishi9 @ddrone